+++ src/base_lib/base64.h
@@ -3,8 +3,8 @@
 
 namespace license {
 
-unsigned char* unbase64(const char* ascii, int len, int *flen);
-char* base64(const void* binaryData, int len, int *flen);
+unsigned char* unbase64(const char *ascii, int len, int *flen);
+char* base64(const void *binaryData, size_t len, size_t *flen);
 
 }
